# bybit_eu_swing_radar/action/openapi.yaml
openapi: 3.1.0
info:
  title: Bybit EU Trading Radar API
  version: 0.7.3
  description: Bybit EU Trading Radar with day-trade strategy v0.7.3, compact audit,
    and context-only derivatives OI/funding Flow feature v0.7.2.2.
servers:
- url: https://bybit-eu-swing-radar-production.up.railway.app
paths:
...
  /v1/scan:
    get:
      operationId: getSwingScan
      summary: Return ranked long and short swing candidates
      parameters:
      - in: query
        name: direction
        schema:
          type: string
          enum:
          - long
          - short
          - both
          default: both
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 10
          default: 3
      - in: query
        name: min_score
        schema:
          type: number
          minimum: 0
          maximum: 100
          default: 70
      responses:
        '200':
          description: Ranked scan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
  /v1/top-candidates:
    get:
      operationId: getTopCandidates
      summary: Use this operation to show the strongest long and short candidates.
        Returns compact strict rankings and separately labelled WATCH_ONLY fallbacks.
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
      - in: query
        name: include_watchlist
        schema:
          type: boolean
          default: true
      responses:
        '200':
          description: Compact strict and watch-only rankings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopCandidatesResponse'
...
  /v1/setup/{symbol}:
    get:
      operationId: getSymbolSetup
      summary: Return detailed bullish and bearish scenarios for one symbol
      parameters:
      - in: path
        name: symbol
        required: true
        schema:
          type: string
        description: Bybit symbol, for example BTCUSDC
      responses:
        '200':
          description: Symbol setup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setup'
        '404':
          description: Symbol not found
...
  /v1/watchlist:
    get:
      operationId: getWatchlist
      summary: Return extended USDC discovery watchlist, including liquidity-blocked
        ideas
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 20
          default: 20
      responses:
        '200':
          description: Extended watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistResponse'
  /v1/momentum-radar:
    get:
      operationId: getMomentumRadar
      summary: Scan every eligible Bybit EU USDC spot pair for early explosive momentum
        and chase risk
      parameters:
      - in: query
        name: direction
        schema:
          type: string
          enum:
          - long
          - short
          - both
          default: both
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 150
          default: 20
      - in: query
        name: min_score
        schema:
          type: number
          minimum: 0
          maximum: 100
          default: 50
      responses:
        '200':
          description: Ranked explosive-move radar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MomentumResponse'
  /v1/day-trade/top-candidates:
    get:
      operationId: getDayTradeTopCandidates
      summary: Show the strongest strict day-trade long and short candidates, with
        separate WATCH_ONLY fallbacks.
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
      - in: query
        name: include_watchlist
        schema:
          type: boolean
          default: true
      responses:
        '200':
          description: Compact day-trade rankings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeTopCandidatesResponse'
  /v1/day-trade/scan:
    get:
      operationId: getDayTradeScan
      summary: Return the cached v0.7.3 day-trade scan using 4H/1H context and
        closed 5m sweep/reclaim/structure confirmation with closed 15m confirmation.
      parameters:
      - in: query
        name: direction
        schema:
          type: string
          enum:
          - long
          - short
          - both
          default: both
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 20
          default: 10
      - in: query
        name: min_score
        schema:
          type: number
          minimum: 0
          maximum: 100
          default: 0
      - in: query
        name: include_watchlist
        schema:
          type: boolean
          default: true
      responses:
        '200':
          description: Day-trade scan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeScanResponse'
  /v1/day-trade/setup/{symbol}:
    get:
      operationId: getDayTradeSetup
      summary: Analyze one cached Bybit EU USDC symbol in day-trade mode.
      parameters:
      - in: path
        name: symbol
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Day-trade setup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeCandidate'
...
  /v1/day-trade/journal/summary:
    get:
      operationId: getDayTradeJournalSummary
      summary: Return prospective day-trade outcome statistics. Use for journal, performance,
        expectancy or edge questions. STRICT and SHADOW samples remain separate.
      parameters:
      - in: query
        name: days
        schema:
          type: integer
          minimum: 1
          maximum: 365
          default: 30
      - in: query
        name: signal_class
        schema:
          type: string
          enum:
          - all
          - STRICT
          - SHADOW
          default: all
      responses:
        '200':
          description: Prospective journal summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeJournalSummaryResponse'
  /v1/day-trade/journal/signals:
    get:
      operationId: getDayTradeJournalSignals
      summary: List prospective day-trade journal signals and their open or closed
        outcomes.
      parameters:
      - in: query
        name: status
        schema:
          type: string
          enum:
          - all
          - OPEN
          - CLOSED
          default: all
      - in: query
        name: signal_class
        schema:
          type: string
          enum:
          - all
          - STRICT
          - SHADOW
          default: all
      - in: query
        name: symbol
        schema:
          type:
          - string
          - 'null'
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
      responses:
        '200':
          description: Journal signal list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeJournalSignalsResponse'
...
  /v1/day-trade/backtest/summary:
    get:
      operationId: getDayTradeBacktestSummary
      summary: Return historical day-trade replay statistics. Never merge these results
        with prospective journal statistics.
      parameters:
      - in: query
        name: signal_class
        schema:
          type: string
          enum:
          - all
          - STRICT
          - SHADOW
          default: all
      - in: query
        name: side
        schema:
          type: string
          enum:
          - both
          - long
          - short
          default: both
      - in: query
        name: primary_only
        schema:
          type: boolean
          default: true
      responses:
        '200':
          description: Backtest summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeBacktestSummaryResponse'
  /v1/day-trade/backtest/signals:
    get:
      operationId: getDayTradeBacktestSignals
      summary: Return individual historical replay signals and outcomes.
      parameters:
      - in: query
        name: signal_class
        schema:
          type: string
          enum:
          - all
          - STRICT
          - SHADOW
          default: all
      - in: query
        name: side
        schema:
          type: string
          enum:
          - both
          - long
          - short
          default: both
      - in: query
        name: symbol
        schema:
          type:
          - string
          - 'null'
      - in: query
        name: primary_only
        schema:
          type: boolean
          default: true
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
      responses:
        '200':
          description: Backtest signals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeBacktestSignalsResponse'
...
  /v1/day-trade/backtest/diagnostics/waterfall:
    get:
      operationId: getDayTradeGateWaterfall
      summary: Show sequential strict gate waterfall and first failures
      parameters:
      - name: side
        in: query
        required: false
        schema:
          type: string
          enum:
          - both
          - long
          - short
          default: both
      - name: split
        in: query
        required: false
        schema:
          type: string
          enum:
          - all
          - DEVELOPMENT
          - VALIDATION
          default: all
      - name: universe_group
        in: query
        required: false
        schema:
          type: string
          enum:
          - all
          - MAJOR_LIQUID
          - OTHER
          default: all
      - name: primary_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Strict gate waterfall
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeGateWaterfallResponse'
  /v1/day-trade/backtest/diagnostics/edge:
    get:
      operationId: getDayTradeEdgeDiagnostics
      summary: Analyze cohort edge, validation split and cost/horizon sensitivity
      parameters:
      - name: cohort
        in: query
        required: false
        schema:
          type: string
          enum:
          - ALL_VALID_CANDIDATES
          - LIQUID_EXECUTABLE
          - SCORE_GATES_PASS
          - NEAR_STRICT
          - STRICT_ELIGIBLE
          - STRICT_TRADE
          default: NEAR_STRICT
      - name: side
        in: query
        required: false
        schema:
          type: string
          enum:
          - both
          - long
          - short
          default: both
      - name: split
        in: query
        required: false
        schema:
          type: string
          enum:
          - all
          - DEVELOPMENT
          - VALIDATION
          default: all
      - name: universe_group
        in: query
        required: false
        schema:
          type: string
          enum:
          - all
          - MAJOR_LIQUID
          - OTHER
          default: all
      - name: primary_only
        in: query
        required: false
        schema:
          type: boolean
          default: true
      responses:
        '200':
          description: Edge diagnostics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeEdgeDiagnosticsResponse'
  /v1/day-trade/audit/{symbol}:
    get:
      operationId: getDayTradeAudit
      summary: Return a compact two-sided day-trade calculation audit for one cached
        USDC symbol.
      description: Use this endpoint for RR, target-path and structural-barrier audits.
        It is intentionally compact to avoid ResponseTooLarge errors from the full
        raw scan.
      parameters:
      - in: path
        name: symbol
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Compact day-trade symbol audit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeSymbolAuditResponse'
...
  /v1/day-trade/flow/{symbol}:
    get:
      operationId: getDayTradeFlowContext
      summary: Return context-only OI/funding flow data for one cached day-trade symbol.
      description: Bybit global linear derivatives plus already-cached Coinalyze secondary
        context. This endpoint does not change v0.7.3 STRICT gates or prove Bybit
        EU execution conditions.
      parameters:
      - in: path
        name: symbol
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Day-trade derivatives flow context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayTradeFlowContextResponse'
components:
  securitySchemes:
    RadarApiKey:
      type: apiKey
      in: header
      name: X-Radar-Key
  schemas:
...
    ScanResponse:
      type: object
      required:
      - data_as_of
      - data_quality
      - market_regime
      - longs
      - shorts
      properties:
        data_as_of:
          type: string
          format: date-time
        data_as_of_budapest:
          type: string
        data_quality:
          type: string
          enum:
          - GOOD
          - PARTIAL
          - DEGRADED
        market_regime:
          $ref: '#/components/schemas/MarketRegime'
        longs:
          type: array
          items:
            $ref: '#/components/schemas/Setup'
        shorts:
          type: array
          items:
            $ref: '#/components/schemas/Setup'
        exclusions:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              reason:
                type: string
        extended_watchlist:
          type: array
          items:
            $ref: '#/components/schemas/Setup'
        liquidity_blocked:
          type: array
          items:
            $ref: '#/components/schemas/Setup'
        universe_stats:
          type: object
          properties:
            raw_usdc_instrument_records:
              type: integer
            unique_usdc_instruments:
              type: integer
            duplicate_instrument_records:
              type: integer
            duplicate_symbols:
              type: array
              items:
                type: string
            active_usdc_pairs:
              type: integer
            eligible_discovery_pairs:
              type: integer
            analysis_universe_size:
              type: integer
            tradeable_pairs:
              type: integer
            liquidity_blocked_pairs:
              type: integer
          additionalProperties: true
        coverage:
          type: object
          properties:
            analyzed_symbols:
              type: integer
            coinalyze_enriched_symbols:
              type: integer
            borrowability_checked_symbols:
              type: integer
            symbols_missing_history:
              type: array
              items:
                type: string
            momentum_eligible_pairs:
              type: integer
            momentum_scanned_pairs:
              type: integer
            momentum_failed_pairs:
              type: integer
            momentum_failed_symbols:
              type: array
              items:
                type: string
            momentum_calculation_failed_pairs:
              type: integer
            momentum_no_activity_pairs:
              type: integer
          additionalProperties: true
        momentum_radar:
          type: array
          items:
            $ref: '#/components/schemas/MomentumItem'
    Setup:
      type: object
      required:
      - symbol
      - side
      - state
      - setup_score
      - expansion_score
      - direction_score
      - quality_score
      - data_as_of
      properties:
        symbol:
          type: string
        base_asset:
          type: string
        quote_asset:
          type: string
        side:
          type: string
          enum:
          - long
          - short
          - neutral
        state:
          type: string
          enum:
          - NO_TRADE
          - WATCH
          - ARMED
          - TRIGGERED
          - MANAGED
          - INVALIDATED
          - EXPIRED
        grade:
          type: string
          enum:
          - A
          - B
          - WATCH
          - NO_TRADE
        confidence:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
        last_price:
          type: number
        shortable:
          type: boolean
        execution_modes:
          type: array
          items:
            type: string
        setup_type:
          type: string
        thesis:
          type: array
          items:
            type: string
        expansion_score:
          type: number
        direction_score:
          type: number
        quality_score:
          type: number
        setup_score:
          type: number
        trigger:
          $ref: '#/components/schemas/PriceCondition'
        entry_zone:
          $ref: '#/components/schemas/PriceZone'
        stop:
          type: number
        invalidation:
          type: string
        targets:
          type: array
          items:
            type: number
        expected_rr:
          type: number
        expected_holding_days:
          type: string
        metrics:
          $ref: '#/components/schemas/SwingMetrics'
        bullish_scenario:
          type: string
        bearish_scenario:
          type: string
        weakest_point:
          type: string
        risks:
          type: array
          items:
            type: string
        data_quality:
          type: string
          enum:
          - GOOD
          - PARTIAL
          - DEGRADED
        missing_data:
          type: array
          items:
            type: string
        data_as_of:
          type: string
          format: date-time
...
    PriceZone:
      type: object
      properties:
        low:
          type: number
        high:
          type: number
...
    WatchlistResponse:
      type: object
      required:
      - data_as_of
      - items
      properties:
        data_as_of:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/Setup'
...
    MomentumResponse:
      type: object
      required:
      - data_as_of
      - scanned_pairs
      - minimum_score
      - items
      properties:
        data_as_of:
          type: string
          format: date-time
        raw_usdc_instrument_records:
          type:
          - integer
          - 'null'
        unique_usdc_instruments:
          type:
          - integer
          - 'null'
        duplicate_instrument_records:
          type: integer
          default: 0
        duplicate_symbols:
          type: array
          items:
            type: string
        eligible_pairs:
          type:
          - integer
          - 'null'
        scanned_pairs:
          type: integer
        failed_pairs:
          type: integer
          default: 0
        failed_symbols:
          type: array
          items:
            type: string
        calculation_failed_pairs:
          type: integer
          default: 0
        calculation_failures:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              reason:
                type: string
        no_activity_pairs:
          type: integer
          default: 0
        minimum_score:
          type: number
        promotion_minimum_score:
          type:
          - number
          - 'null'
        items:
          type: array
          items:
            $ref: '#/components/schemas/MomentumItem'
...
    TopCandidatesResponse:
      type: object
      required:
      - data_as_of
      - data_as_of_budapest
      - data_quality
      - requested_limit
      - strict_long_count
      - strict_short_count
      - strict_longs
      - strict_shorts
      - watch_only_longs
      - watch_only_shorts
      properties:
        data_as_of:
          type: string
          format: date-time
        data_as_of_budapest:
          type: string
        data_quality:
          type: string
          enum:
          - GOOD
          - PARTIAL
          - DEGRADED
        market_regime:
          $ref: '#/components/schemas/CompactMarketRegime'
        requested_limit:
          type: integer
        strict_long_count:
          type: integer
        strict_short_count:
          type: integer
        strict_longs:
          type: array
          items:
            $ref: '#/components/schemas/CompactCandidate'
        strict_shorts:
          type: array
          items:
            $ref: '#/components/schemas/CompactCandidate'
        watch_only_longs:
          type: array
          items:
            $ref: '#/components/schemas/CompactCandidate'
        watch_only_shorts:
          type: array
          items:
            $ref: '#/components/schemas/CompactCandidate'
        notes:
          type: array
          items:
            type: string
...
    DayTradeTopCandidatesResponse:
      allOf:
      - $ref: '#/components/schemas/DayTradeScanResponse'
      - type: object
        required:
        - requested_limit
        - strict_long_count
        - strict_short_count
        properties:
          requested_limit:
            type: integer
          strict_long_count:
            type: integer
          strict_short_count:
            type: integer
...
    DayTradeJournalSummaryResponse:
      type: object
      required:
      - strategy_version
      - generated_at
      - window_days
      - requested_signal_class
      - evidence_status
      - strict_closed_sample
      - overall
      - methodology
      - warnings
      properties:
        strategy_version:
          type: string
        generated_at:
          type: string
          format: date-time
        window_days:
          type: integer
        requested_signal_class:
          type: string
          enum:
          - all
          - STRICT
          - SHADOW
        evidence_status:
          type: string
          enum:
          - INSUFFICIENT_SAMPLE
          - EARLY_SAMPLE
          - EVALUABLE_SAMPLE
        strict_closed_sample:
          type: integer
        overall:
          $ref: '#/components/schemas/JournalAggregate'
        by_signal_class:
          type: array
          items:
            $ref: '#/components/schemas/JournalGroupStats'
        by_side:
          type: array
          items:
            $ref: '#/components/schemas/JournalGroupStats'
        by_setup_type:
          type: array
          items:
            $ref: '#/components/schemas/JournalGroupStats'
        latest_run:
          $ref: '#/components/schemas/JournalLatestRun'
        methodology:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
      additionalProperties: false
...
security:
- RadarApiKey: []